+++ Scripts/Laura Stroop.py
@@ -14,6 +14,7 @@ import psyneulink as pnl
 
 # SET UP MECHANISMS
 #   Linear input units, colors: ('red', 'green'), words: ('RED','GREEN')
+import psyneulink.core.components.functions.distributionfunctions
 import psyneulink.core.components.functions.integratorfunctions
 import psyneulink.core.components.functions.transferfunctions
 
@@ -34,7 +35,7 @@ colors_hidden_layer = pnl.TransferMechanism(size=2,
                                             function=psyneulink.core.components.functions.transferfunctions.Logistic(gain=1.0, x_0=4.0),
                                             # function=pnl.Logistic(gain=1.0, offset=-4.0),
                                             integrator_mode=False,
-                                            noise=pnl.NormalDist(mean=0.0, standard_deviation=.01).function,
+                                            noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0.0, standard_deviation=.01).function,
                                             integration_rate=0.1,
                                             name='COLORS HIDDEN')
 #should be randomly distributed noise to the net input of each unit (except input unit)
@@ -43,7 +44,7 @@ words_hidden_layer = pnl.TransferMechanism(size=2,
                                            function=psyneulink.core.components.functions.transferfunctions.Logistic(gain=1.0, x_0=4.0),
                                            # function=pnl.Logistic(gain=1.0, offset=-4.0),
                                            integrator_mode=False,
-                                           noise=pnl.NormalDist(mean=0.0, standard_deviation=.01).function,
+                                           noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0.0, standard_deviation=.01).function,
                                            integration_rate=0.1,
                                            name='WORDS HIDDEN')
 
@@ -66,7 +67,7 @@ response_layer = pnl.TransferMechanism(size=2,
                                        function=psyneulink.core.components.functions.transferfunctions.Logistic,
                                        name='RESPONSE',
                                        integrator_mode=True,
-                                       noise=pnl.NormalDist(mean=0.0, standard_deviation=.01).function,
+                                       noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0.0, standard_deviation=.01).function,
                                        integration_rate=0.1)
 #   Respond red accumulator
 #parameters from paper
