+++ psyneulink/core/components/functions/__init__.py
@@ -5,6 +5,7 @@ from . import interfacefunctions
 from . import transferfunctions
 from . import selectionfunctions
 from . import integratorfunctions
+from . import distributionfunctions
 
 from .function import *
 from .userdefinedfunction import *
@@ -13,6 +14,7 @@ from .interfacefunctions import *
 from .transferfunctions import *
 from .selectionfunctions import *
 from .integratorfunctions import *
+from .distributionfunctions import *
 
 __all__ = list(function.__all__)
 __all__.extend(userdefinedfunction.__all__)
@@ -21,3 +23,4 @@ __all__.extend(interfacefunctions.__all__)
 __all__.extend(transferfunctions.__all__)
 __all__.extend(selectionfunctions.__all__)
 __all__.extend(integratorfunctions.__all__)
+__all__.extend(distributionfunctions.__all__)
