+++ psyneulink/core/components/functions/function.py
@@ -16,14 +16,6 @@ Function
 Example function:
   * `ArgumentTherapy`
 
-Distribution Functions:
-  * `NormalDist`
-  * `UniformToNormalDist`
-  * `ExponentialDist`
-  * `UniformDist`
-  * `GammaDist`
-  * `WaldDist`
-
 Objective Functions:
   * `Stability`
   * `Distance`
@@ -183,16 +175,14 @@ from psyneulink.core.components.functions.transferfunctions import Logistic, get
 from psyneulink.core.components.shellclasses import Function, Mechanism
 from psyneulink.core.globals.context import ContextFlags
 from psyneulink.core.globals.defaults import MPI_IMPLEMENTATION
-from psyneulink.core.globals.keywords import ARGUMENT_THERAPY_FUNCTION, BACKPROPAGATION_FUNCTION, BETA, \
-    CONTRASTIVE_HEBBIAN_FUNCTION, CORRELATION, CROSS_ENTROPY, DEFAULT_VARIABLE, DIFFERENCE, DISTANCE_FUNCTION, DISTANCE_METRICS, DIST_FUNCTION_TYPE, DIST_MEAN, DIST_SHAPE, \
-    DistanceMetrics, ENERGY, ENTROPY, EUCLIDEAN, EXAMPLE_FUNCTION_TYPE, EXPONENTIAL, EXPONENTIAL_DIST_FUNCTION, \
-    FUNCTION, FUNCTION_OUTPUT_TYPE, FUNCTION_OUTPUT_TYPE_CONVERSION, GAMMA_DIST_FUNCTION, GAUSSIAN, \
-    GRADIENT_OPTIMIZATION_FUNCTION, GRID_SEARCH_FUNCTION, HEBBIAN_FUNCTION, HIGH, HOLLOW_MATRIX, \
+from psyneulink.core.globals.keywords import ARGUMENT_THERAPY_FUNCTION, BACKPROPAGATION_FUNCTION, \
+    CONTRASTIVE_HEBBIAN_FUNCTION, CORRELATION, CROSS_ENTROPY, DEFAULT_VARIABLE, DIFFERENCE, DISTANCE_FUNCTION, DISTANCE_METRICS, \
+    DistanceMetrics, ENERGY, ENTROPY, EUCLIDEAN, EXAMPLE_FUNCTION_TYPE, EXPONENTIAL, FUNCTION, FUNCTION_OUTPUT_TYPE, FUNCTION_OUTPUT_TYPE_CONVERSION, \
+    GAUSSIAN, \
+    GRADIENT_OPTIMIZATION_FUNCTION, GRID_SEARCH_FUNCTION, HEBBIAN_FUNCTION, HOLLOW_MATRIX, \
     KOHONEN_FUNCTION, LEARNING_FUNCTION_TYPE, LEARNING_RATE, LINEAR, \
-    LOW, MATRIX, MAX_ABS_DIFF, NAME, NORMAL_DIST_FUNCTION, OBJECTIVE_FUNCTION_TYPE, OPTIMIZATION_FUNCTION_TYPE, OWNER, PARAMETER_STATE_PARAMS, \
-    RL_FUNCTION, SCALE, STABILITY_FUNCTION, STANDARD_DEVIATION, \
-    TDLEARNING_FUNCTION, UNIFORM_DIST_FUNCTION, VALUE, VARIABLE, \
-    WALD_DIST_FUNCTION, \
+    MATRIX, MAX_ABS_DIFF, NAME, OBJECTIVE_FUNCTION_TYPE, OPTIMIZATION_FUNCTION_TYPE, OWNER, PARAMETER_STATE_PARAMS, \
+    RL_FUNCTION, STABILITY_FUNCTION, TDLEARNING_FUNCTION, VALUE, VARIABLE, \
     kwComponentCategory, kwPreferenceSetName
 from psyneulink.core.globals.parameters import Param
 from psyneulink.core.globals.preferences.componentpreferenceset import is_pref_set, kpReportOutputPref
@@ -207,18 +197,19 @@ __all__ = [
     'ADDITIVE', 'ADDITIVE_PARAM', 'ASCENT',
     'AdditiveParam', 'ArgumentTherapy', 'AUTOASSOCIATIVE',
     'BackPropagation', 'BayesGLM', 'ContrastiveHebbian',
-    'DESCENT', 'DISABLE', 'DISABLE_PARAM', 'Distance', 'DistributionFunction', 'EPSILON',
-    'ERROR_MATRIX', 'ExponentialDist', 'Function_Base', 'function_keywords', 'FunctionError', 'FunctionOutputType', 'FunctionRegistry',
-    'GammaDist', 'get_param_value_for_function', 'get_param_value_for_keyword',
+    'DESCENT', 'DISABLE', 'DISABLE_PARAM', 'Distance', 'EPSILON',
+    'ERROR_MATRIX', 'Function_Base', 'function_keywords', 'FunctionError', 'FunctionOutputType', 'FunctionRegistry',
+    'get_param_value_for_function', 'get_param_value_for_keyword',
     'GradientOptimization', 'GridSearch',
     'Hebbian', 'is_Function', 'is_function_type',
     'LEARNING_ACTIVATION_FUNCTION',
     'LEARNING_ACTIVATION_INPUT', 'LEARNING_ACTIVATION_OUTPUT',
     'LEARNING_ERROR_OUTPUT', 'LearningFunction', 'MAXIMIZE', 'max_vs_avg', 'max_vs_next', 'MINIMIZE', 'ModulatedParam',
     'ModulationParam', 'MULTIPLICATIVE', 'MULTIPLICATIVE_PARAM',
-    'MultiplicativeParam', 'NormalDist', 'ObjectiveFunction', 'OptimizationFunction', 'OVERRIDE', 'OVERRIDE_PARAM', 'PERTINACITY', 'PROPENSITY',
+    'MultiplicativeParam', 'ObjectiveFunction', 'OptimizationFunction', 'OVERRIDE', 'OVERRIDE_PARAM', 'PERTINACITY', 'PROPENSITY',
     'Reinforcement', 'ReturnVal',
-    'SEARCH_FUNCTION', 'SEARCH_SPACE', 'SEARCH_TERMINATION_FUNCTION', 'Stability', 'TDLearning', 'UniformDist', 'UniformToNormalDist', 'WaldDist', 'WT_MATRIX_RECEIVERS_DIM', 'WT_MATRIX_SENDERS_DIM'
+    'SEARCH_FUNCTION', 'SEARCH_SPACE', 'SEARCH_TERMINATION_FUNCTION', 'Stability', 'TDLearning',
+    'WT_MATRIX_RECEIVERS_DIM', 'WT_MATRIX_SENDERS_DIM'
 ]
 
 
@@ -1132,662 +1123,6 @@ class ArgumentTherapy(Function_Base):
 
 
 # region ****************************************   FUNCTIONS   ********************************************************
-# endregion
-
-# region ************************************   DISTRIBUTION FUNCTIONS   ***********************************************
-
-class DistributionFunction(Function_Base):
-    componentType = DIST_FUNCTION_TYPE
-
-
-class NormalDist(DistributionFunction):
-    """
-    NormalDist(                      \
-             mean=0.0,               \
-             standard_deviation=1.0, \
-             params=None,            \
-             owner=None,             \
-             prefs=None              \
-             )
-
-    .. _NormalDist:
-
-    Return a random sample from a normal distribution using numpy.random.normal
-
-    Arguments
-    ---------
-
-    mean : float : default 0.0
-        The mean or center of the normal distribution
-
-    standard_deviation : float : default 1.0
-        Standard deviation of the normal distribution. Must be > 0.0
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    mean : float : default 0.0
-        The mean or center of the normal distribution
-
-    standard_deviation : float : default 1.0
-        Standard deviation of the normal distribution. Must be > 0.0
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    """
-
-    componentName = NORMAL_DIST_FUNCTION
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    class Params(DistributionFunction.Params):
-        mean = Param(0.0, modulable=True)
-        standard_deviation = Param(1.0, modulable=True)
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 mean=0.0,
-                 standard_deviation=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(mean=mean,
-                                                  standard_deviation=standard_deviation,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def _validate_params(self, request_set, target_set=None, context=None):
-        super()._validate_params(request_set=request_set, target_set=target_set, context=context)
-
-        if STANDARD_DEVIATION in target_set:
-            if target_set[STANDARD_DEVIATION] <= 0.0:
-                raise FunctionError("The standard_deviation parameter ({}) of {} must be greater than zero.".
-                                    format(target_set[STANDARD_DEVIATION], self.name))
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        mean = self.get_current_function_param(DIST_MEAN, execution_id)
-        standard_deviation = self.get_current_function_param(STANDARD_DEVIATION, execution_id)
-
-        result = np.random.normal(mean, standard_deviation)
-
-        return self.convert_output_type(result)
-
-
-class UniformToNormalDist(DistributionFunction):
-    """
-    UniformToNormalDist(             \
-             mean=0.0,               \
-             standard_deviation=1.0, \
-             params=None,            \
-             owner=None,             \
-             prefs=None              \
-             )
-
-    .. _UniformToNormalDist:
-
-    Return a random sample from a normal distribution using first np.random.rand(1) to generate a sample from a uniform
-    distribution, and then converting that sample to a sample from a normal distribution with the following equation:
-
-    .. math::
-
-        normal\\_sample = \\sqrt{2} \\cdot standard\\_dev \\cdot scipy.special.erfinv(2 \\cdot uniform\\_sample - 1)  + mean
-
-    The uniform --> normal conversion allows for a more direct comparison with MATLAB scripts.
-
-    .. note::
-
-        This function requires `SciPy <https://pypi.python.org/pypi/scipy>`_.
-
-    (https://github.com/jonasrauber/randn-matlab-python)
-
-    Arguments
-    ---------
-
-    mean : float : default 0.0
-        The mean or center of the normal distribution
-
-    standard_deviation : float : default 1.0
-        Standard deviation of the normal distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    mean : float : default 0.0
-        The mean or center of the normal distribution
-
-    standard_deviation : float : default 1.0
-        Standard deviation of the normal distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    """
-
-    componentName = NORMAL_DIST_FUNCTION
-
-    class Params(DistributionFunction.Params):
-        variable = Param(np.array([0]), read_only=True)
-        mean = Param(0.0, modulable=True)
-        standard_deviation = Param(1.0, modulable=True)
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 mean=0.0,
-                 standard_deviation=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(mean=mean,
-                                                  standard_deviation=standard_deviation,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-
-        try:
-            from scipy.special import erfinv
-        except:
-            raise FunctionError("The UniformToNormalDist function requires the SciPy package.")
-
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        mean = self.get_current_function_param(DIST_MEAN, execution_id)
-        standard_deviation = self.get_current_function_param(STANDARD_DEVIATION, execution_id)
-
-        sample = np.random.rand(1)[0]
-        result = ((np.sqrt(2) * erfinv(2 * sample - 1)) * standard_deviation) + mean
-
-        return self.convert_output_type(result)
-
-
-class ExponentialDist(DistributionFunction):
-    """
-    ExponentialDist(                \
-             beta=1.0,              \
-             params=None,           \
-             owner=None,            \
-             prefs=None             \
-             )
-
-    .. _ExponentialDist:
-
-    Return a random sample from a exponential distribution using numpy.random.exponential
-
-    Arguments
-    ---------
-
-    beta : float : default 1.0
-        The scale parameter of the exponential distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    beta : float : default 1.0
-        The scale parameter of the exponential distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    """
-    componentName = EXPONENTIAL_DIST_FUNCTION
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    class Params(DistributionFunction.Params):
-        beta = Param(1.0, modulable=True)
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 beta=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(beta=beta,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        beta = self.get_current_function_param(BETA, execution_id)
-        result = np.random.exponential(beta)
-
-        return self.convert_output_type(result)
-
-
-class UniformDist(DistributionFunction):
-    """
-    UniformDist(                      \
-             low=0.0,             \
-             high=1.0,             \
-             params=None,           \
-             owner=None,            \
-             prefs=None             \
-             )
-
-    .. _UniformDist:
-
-    Return a random sample from a uniform distribution using numpy.random.uniform
-
-    Arguments
-    ---------
-
-    low : float : default 0.0
-        Lower bound of the uniform distribution
-
-    high : float : default 1.0
-        Upper bound of the uniform distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    low : float : default 0.0
-        Lower bound of the uniform distribution
-
-    high : float : default 1.0
-        Upper bound of the uniform distribution
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    """
-    componentName = UNIFORM_DIST_FUNCTION
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    class Params(DistributionFunction.Params):
-        low = Param(0.0, modulable=True)
-        high = Param(1.0, modulable=True)
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 low=0.0,
-                 high=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(low=low,
-                                                  high=high,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        low = self.get_current_function_param(LOW, execution_id)
-        high = self.get_current_function_param(HIGH, execution_id)
-        result = np.random.uniform(low, high)
-
-        return self.convert_output_type(result)
-
-
-class GammaDist(DistributionFunction):
-    """
-    GammaDist(\
-             scale=1.0,\
-             dist_shape=1.0,\
-             params=None,\
-             owner=None,\
-             prefs=None\
-             )
-
-    .. _GammaDist:
-
-    Return a random sample from a gamma distribution using numpy.random.gamma
-
-    Arguments
-    ---------
-
-    scale : float : default 1.0
-        The scale of the gamma distribution. Should be greater than zero.
-
-    dist_shape : float : default 1.0
-        The shape of the gamma distribution. Should be greater than zero.
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    scale : float : default 1.0
-        The dist_shape of the gamma distribution. Should be greater than zero.
-
-    dist_shape : float : default 1.0
-        The scale of the gamma distribution. Should be greater than zero.
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    """
-
-    componentName = GAMMA_DIST_FUNCTION
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    class Params(DistributionFunction.Params):
-        scale = Param(1.0, modulable=True)
-        dist_shape = Param(1.0, modulable=True)
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 scale=1.0,
-                 dist_shape=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(scale=scale,
-                                                  dist_shape=dist_shape,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        scale = self.get_current_function_param(SCALE, execution_id)
-        dist_shape = self.get_current_function_param(DIST_SHAPE, execution_id)
-
-        result = np.random.gamma(dist_shape, scale)
-
-        return self.convert_output_type(result)
-
-
-class WaldDist(DistributionFunction):
-    """
-     WaldDist(             \
-              scale=1.0,\
-              mean=1.0,\
-              params=None,\
-              owner=None,\
-              prefs=None\
-              )
-
-     .. _WaldDist:
-
-     Return a random sample from a Wald distribution using numpy.random.wald
-
-     Arguments
-     ---------
-
-     scale : float : default 1.0
-         Scale parameter of the Wald distribution. Should be greater than zero.
-
-     mean : float : default 1.0
-         Mean of the Wald distribution. Should be greater than or equal to zero.
-
-     params : Dict[param keyword: param value] : default None
-         a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-         function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-         arguments of the constructor.
-
-     owner : Component
-         `component <Component>` to which to assign the Function.
-
-     prefs : PreferenceSet or specification dict : default Function.classPreferences
-         the `PreferenceSet` for the Function. If it is not specified, a default is assigned using `classPreferences`
-         defined in __init__.py (see :doc:`PreferenceSet <LINK>` for details).
-
-
-     Attributes
-     ----------
-
-     scale : float : default 1.0
-         Scale parameter of the Wald distribution. Should be greater than zero.
-
-     mean : float : default 1.0
-         Mean of the Wald distribution. Should be greater than or equal to zero.
-
-     params : Dict[param keyword: param value] : default None
-         a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-         function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-         arguments of the constructor.
-
-     owner : Component
-         `component <Component>` to which to assign the Function.
-
-     prefs : PreferenceSet or specification dict : default Function.classPreferences
-         the `PreferenceSet` for the Function. If it is not specified, a default is assigned using `classPreferences`
-         defined in __init__.py (see :doc:`PreferenceSet <LINK>` for details).
-
-
-     """
-
-    componentName = WALD_DIST_FUNCTION
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-
-    class Params(DistributionFunction.Params):
-        scale = Param(1.0, modulable=True)
-        mean = Param(1.0, modulable=True)
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 scale=1.0,
-                 mean=1.0,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(scale=scale,
-                                                  mean=mean,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-
-    def function(self,
-                 variable=None,
-                 execution_id=None,
-                 params=None,
-                 context=None):
-        # Validate variable and validate params
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        scale = self.get_current_function_param(SCALE, execution_id)
-        mean = self.get_current_function_param(DIST_MEAN, execution_id)
-
-        result = np.random.wald(mean, scale)
-
-        return self.convert_output_type(result)
-
-
 # endregion
 
 # region **************************************  OBJECTIVE FUNCTIONS ***************************************************
