+++ tests/composition/test_models.py
@@ -2,6 +2,7 @@ import numpy as np
 import psyneulink as pnl
 import pytest
 
+import psyneulink.core.components.functions.distributionfunctions
 import psyneulink.core.components.functions.integratorfunctions
 import psyneulink.core.components.functions.transferfunctions
 import psyneulink.core.globals.utilities
@@ -314,14 +315,15 @@ class TestModels:
                                                     function=psyneulink.core.components.functions.transferfunctions.Logistic(gain=1.0, x_0=4.0),
                                                     # should be able to get same result with offset = -4.0
                                                     integrator_mode=True,
-                                                    noise=pnl.NormalDist(mean=0, standard_deviation=unit_noise).function,
+                                                    noise=psyneulink.core.components.functions.distributionfunctions
+                                                    .NormalDist(mean=0, standard_deviation=unit_noise).function,
                                                     integration_rate=0.1,
                                                     name='COLORS HIDDEN')
         #    words_hidden: ('RED','GREEN')
         words_hidden_layer = pnl.TransferMechanism(size=2,
                                                    function=psyneulink.core.components.functions.transferfunctions.Logistic(gain=1.0, x_0=4.0),
                                                    integrator_mode=True,
-                                                   noise=pnl.NormalDist(mean=0, standard_deviation=unit_noise).function,
+                                                   noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0, standard_deviation=unit_noise).function,
                                                    integration_rate=0.1,
                                                    name='WORDS HIDDEN')
 
@@ -334,7 +336,7 @@ class TestModels:
                                                function=psyneulink.core.components.functions.transferfunctions.Logistic,
                                                name='RESPONSE',
                                                integrator_mode=True,
-                                               noise=pnl.NormalDist(mean=0, standard_deviation=unit_noise).function,
+                                               noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0, standard_deviation=unit_noise).function,
                                                integration_rate=0.1)
         #   Respond red accumulator
         #   alpha = rate of evidence accumlation = 0.1
@@ -342,13 +344,14 @@ class TestModels:
         #   noise will be: squareroot(time_step_size * noise) * a random sample from a normal distribution
         accumulator_noise = 0.1
         respond_red_accumulator = pnl.IntegratorMechanism(function=psyneulink.core.components.functions
-                                                          .integratorfunctions.SimpleIntegrator(noise=pnl.NormalDist(mean=0,
-                                                                                                                                                                  standard_deviation=accumulator_noise).function,
+                                                          .integratorfunctions.SimpleIntegrator(noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0,
+                                                                                                                                                                            standard_deviation=accumulator_noise).function,
                                                                                                                                              rate=0.1),
                                                           name='respond_red_accumulator')
         #   Respond green accumulator
-        respond_green_accumulator = pnl.IntegratorMechanism(function=psyneulink.core.components.functions.integratorfunctions.SimpleIntegrator(noise=pnl.NormalDist(mean=0,
-                                                                                                                                                                    standard_deviation=accumulator_noise).function,
+        respond_green_accumulator = pnl.IntegratorMechanism(function=psyneulink.core.components.functions
+                                                            .integratorfunctions.SimpleIntegrator(noise=psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0,
+                                                                                                                                                                                                                           standard_deviation=accumulator_noise).function,
                                                                                                                                                rate=0.1),
                                                             name='respond_green_accumulator')
 
@@ -505,7 +508,7 @@ class TestModels:
             # Turn on accumulation
             switch_integrator_mode(mechanisms, True)
             # Turn on noise
-            switch_noise(mechanisms, pnl.NormalDist(mean=0, standard_deviation=unit_noise).function)
+            switch_noise(mechanisms, psyneulink.core.components.functions.distributionfunctions.NormalDist(mean=0, standard_deviation=unit_noise).function)
             # Execute until one of the accumulators crosses the threshold
             my_Stroop.termination_processing = {pnl.TimeScale.TRIAL: pnl.While(pass_threshold,
                                                                                respond_red_accumulator,
