+++ tests/mechanisms/test_processing_mechanism.py
@@ -1,8 +1,11 @@
 import numpy as np
 import pytest
 
-from psyneulink.core.components.functions.function import BackPropagation, Distance, ExponentialDist, FunctionError, GammaDist, Hebbian, \
-    NormalDist, Reinforcement, Stability, TDLearning, UniformDist, UniformToNormalDist, WaldDist
+from psyneulink.core.components.functions.function import BackPropagation, Distance, FunctionError, Hebbian, \
+    Reinforcement, Stability, TDLearning
+from psyneulink.core.components.functions.distributionfunctions import NormalDist, UniformToNormalDist, \
+    ExponentialDist, \
+    UniformDist, GammaDist, WaldDist
 from psyneulink.core.components.functions.integratorfunctions import Integrator, SimpleIntegrator, ConstantIntegrator, \
     AdaptiveIntegrator, DriftDiffusionIntegrator, OrnsteinUhlenbeckIntegrator, FHNIntegrator, AccumulatorIntegrator, \
     AGTUtilityIntegrator, BogaczEtAl, NavarroAndFuss
